TRASFERIMENTO CLASSI IN APPCODE
preghiamo
