+++ GENUNISOLUTION/GENUNISOLUTION.sln
@@ -47,8 +47,6 @@ Project("{E24C65DC-7377-472B-9ABA-BC803B73C61A}") = "WEBSERVICE", "WEBSERVICE\",
 		DefaultWebSiteLanguage = "Visual C#"
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LIBCLASSI", "LIBCLASS\LIBCLASSI.csproj", "{DD2273E5-F02F-4787-9589-C61428034B9E}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -63,10 +61,6 @@ Global
 		{A6A37593-0E6A-423A-BA4F-FB4ED73D2AB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A6A37593-0E6A-423A-BA4F-FB4ED73D2AB8}.Release|Any CPU.ActiveCfg = Debug|Any CPU
 		{A6A37593-0E6A-423A-BA4F-FB4ED73D2AB8}.Release|Any CPU.Build.0 = Debug|Any CPU
-		{DD2273E5-F02F-4787-9589-C61428034B9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DD2273E5-F02F-4787-9589-C61428034B9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DD2273E5-F02F-4787-9589-C61428034B9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DD2273E5-F02F-4787-9589-C61428034B9E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
