--- GENUNISOLUTION/LIBRERIACLASSI/RISPOSTE.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LIBCLASSI
-{
-    public class RISPOSTE
-    {
-        public int CHIAVE;
-        public int COD_STUDENTE;
-        public int COD_TEST;
-        public string RISPOSTA;
-
-    }
-}
